+++ prime-router/src/main/kotlin/cli/TestReportStream.kt
@@ -10,30 +10,24 @@ import com.github.ajalt.clikt.parameters.options.flag
 import com.github.ajalt.clikt.parameters.options.option
 import com.github.ajalt.clikt.parameters.options.validate
 import com.github.ajalt.clikt.parameters.types.choice
-import gov.cdc.prime.router.FakeReport
 import gov.cdc.prime.router.FileSettings
-import gov.cdc.prime.router.FileSource
 import gov.cdc.prime.router.Metadata
 import gov.cdc.prime.router.REPORT_MAX_ITEMS
 import gov.cdc.prime.router.REPORT_MAX_ITEM_COLUMNS
 import gov.cdc.prime.router.Receiver
 import gov.cdc.prime.router.Report
 import gov.cdc.prime.router.ReportId
-import gov.cdc.prime.router.Sender
 import gov.cdc.prime.router.azure.DataAccessTransaction
 import gov.cdc.prime.router.azure.DatabaseAccess
+import gov.cdc.prime.router.azure.HttpUtilities
 import gov.cdc.prime.router.azure.ReportFunction
+import gov.cdc.prime.router.azure.ReportStreamEnv
 import gov.cdc.prime.router.azure.WorkflowEngine
 import gov.cdc.prime.router.azure.db.enums.TaskAction
-import gov.cdc.prime.router.serializers.CsvSerializer
-import gov.cdc.prime.router.serializers.Hl7Serializer
-import gov.cdc.prime.router.serializers.RedoxSerializer
 import org.jooq.impl.DSL
 import java.io.File
-import java.io.IOException
 import java.lang.Thread.sleep
 import java.net.HttpURLConnection
-import java.net.URL
 import java.time.OffsetDateTime
 import kotlin.system.exitProcess
 
@@ -41,20 +35,23 @@ class TestReportStream : CliktCommand(
     name = "test",
     help = """Run tests of the Router functions
 
-Database connection info is supplied by environment variables, for localhost, and Test, resp.:
-
+Database connection info is supplied by environment variables.
+Examples for localhost, and Azure Staging, respectively:
+```
 export POSTGRES_USER=prime
 export POSTGRES_URL=jdbc:postgresql://localhost:5432/prime_data_hub
 export POSTGRES_PASSWORD=<secret>
 
-export POSTGRES_USER=prime@pdhtest-pgsql
-export POSTGRES_URL=jdbc:postgresql://pdhtest-pgsql.postgres.database.azure.com:5432/prime_data_hub
+export POSTGRES_USER=prime@pdhstaging-pgsql
+export POSTGRES_URL=jdbc:postgresql://pdhstaging-pgsql.postgres.database.azure.com:5432/prime_data_hub
 export POSTGRES_PASSWORD=<SECRET>
-
+```
 Examples:
-  ./prime test    This runs all the tests locally.
-  
-./prime test --run ping,end2end --env test --key xxxxxxx  This runs the ping and end2end tests in azure Test
+```
+ ./prime test --list        List detailed information about available tests
+ ./prime test               Runs all the tests on localhost
+ ./prime test --run ping,end2end --env staging --key xxxxxxx       Runs the ping and end2end tests in azure Staging
+```
 
 """,
 ) {
@@ -63,20 +60,13 @@ Examples:
     enum class AwesomeTest(val description: String) {
         ping("Is the reports endpoint alive and listening?"),
         end2end("Create Fake data, submit, wait, confirm sent via database lineage data"),
-        strac("Submit data in the strac schema format, wait, confirm via database queries"),
-        stracbasic("Basic strac test to REDOX only."),
-        // 10,000 lines fake data generation took about 90 seconds on my laptop.  6Meg.
+        merge("Submit multiple files, wait, confirm via db that merge occurred"),
+        // Note: 10,000 lines fake data generation took about 90 seconds on my laptop.  6Meg.
         huge("Submit $REPORT_MAX_ITEMS line csv file, wait, confirm via db.  Slow."),
         toobig("Submit ${REPORT_MAX_ITEMS + 1} lines, which should be an error.  Slower ;)"),
         toomanycols("Submit a file with more than $REPORT_MAX_ITEM_COLUMNS columns, which should error"),
-        merge("Submit multiple files, wait, confirm via db that merge occurred"),
-    }
-
-    enum class TestingEnvironment(val endPoint: String) {
-        // track headers and parameters separate from the base endpoint, since they vary
-        TEST("https://pdhtest-functionapp.azurewebsites.net/api/reports"),
-        LOCAL("http://localhost:7071/api/reports"),
-        STAGING("https://pdhstaging-functionapp.azurewebsites.net/api/reports")
+        stracbasic("Basic strac test to REDOX only."),
+        strac("Submit data in the strac schema format, wait, confirm via database queries"),
     }
 
     private val dir by option(
@@ -105,12 +95,13 @@ Examples:
 
     private val env by option(
         "--env",
-        help = "Specify 'local, 'test', or 'staging'.  'local' will connect to ${TestingEnvironment.LOCAL.endPoint}," +
-            " and 'test' will connect to ${TestingEnvironment.TEST.endPoint}"
+        help = "Specify 'local, 'test', or 'staging'.  'local' will connect to ${ReportStreamEnv.LOCAL.endPoint}," +
+            " and 'test' will connect to ${ReportStreamEnv.TEST.endPoint}"
     ).choice("test", "local", "staging").default("local").validate {
         when (it) {
             "test" -> require(!key.isNullOrBlank()) { "Must specify --key <secret> to submit reports to --env test" }
-            "staging" -> require(!key.isNullOrBlank()) { "Must specify --key <secret> to submit reports to --env test" }
+            "staging" ->
+                require(!key.isNullOrBlank()) { "Must specify --key <secret> to submit reports to --env staging" }
         }
     }
 
@@ -141,203 +132,29 @@ Examples:
     }
 
     private fun doTests(tests: List<AwesomeTest>) {
-        val environment = TestingEnvironment.valueOf(env.toUpperCase())
+        val environment = ReportStreamEnv.valueOf(env.toUpperCase())
 
         tests.forEach { test ->
             when (test) {
                 AwesomeTest.ping -> doCheckConnections(environment)
                 AwesomeTest.end2end -> doEndToEndTest(environment)
-                AwesomeTest.strac -> doStracTest(environment)
-                AwesomeTest.stracbasic -> doStracBasicTest(environment)
+                AwesomeTest.merge -> doMergeTest(environment)
                 AwesomeTest.huge -> doHugeTest(environment)
                 AwesomeTest.toobig -> doTooManyItemsTest(environment)
                 AwesomeTest.toomanycols -> doTooManyColumnsTest(environment)
-                AwesomeTest.merge -> doMergeTest(environment)
+                AwesomeTest.stracbasic -> doStracBasicTest(environment)
+                AwesomeTest.strac -> doStracTest(environment)
                 else -> bad("Test $test not implemented")
             }
         }
     }
 
-    private fun doMergeTest(environment: TestingEnvironment) {
-        val fakeItemCount = 20 // hack:  you need use a multiple of # of targetCounties
-        echo("Merge test of: ${environment.endPoint} across $fourTargetCounties")
-        val file = createFakeFile(
-            metadata,
-            simpleRepSender,
-            fakeItemCount,
-            receivingStates,
-            fourTargetCounties,
-            dir,
-        )
-        echo("Created datafile $file")
-        // Now send it to the Hub over and over
-        val reportIds = (1..5).map {
-            val (responseCode, json) =
-                postReportFile(environment, file, org.name, simpleRepSender.name, key)
-            echo("Response to POST: $responseCode")
-            if (responseCode != HttpURLConnection.HTTP_CREATED) {
-                bad("***Merge Test FAILED***:  response code $responseCode")
-                return
-            }
-            val tree = jacksonObjectMapper().readTree(json)
-            val reportId = ReportId.fromString(tree["id"].textValue())
-            echo("Id of submitted report: $reportId")
-            reportId
-        }
-        waitABit(30, environment)
-        examineMergeResults(reportIds[0], fourReceivers, fakeItemCount, 5)
-    }
-
-    private fun doStracTest(
-        environment: TestingEnvironment
-    ) {
-        echo("Test sending Strac data to: ${environment.endPoint} across $fourTargetCounties")
-        val fakeItemCount = 20 // hack:  you need use a multiple of # of targetCounties
-        val file = createFakeFile(
-            metadata,
-            stracSender,
-            fakeItemCount,
-            receivingStates,
-            fourTargetCounties,
-            dir,
-        )
-        echo("Created datafile $file")
-        // Now send it to the Hub.
-        val (responseCode, json) = postReportFile(environment, file, org.name, stracSender.name, key)
-        echo("Response to POST: $responseCode")
-        echo(json)
-        if (responseCode != HttpURLConnection.HTTP_CREATED) {
-            bad("**Strac Test FAILED***:  response code $responseCode")
-            return
-        }
-        val tree = jacksonObjectMapper().readTree(json)
-        val reportId = ReportId.fromString(tree["id"].textValue())
-        echo("Id of submitted report: $reportId")
-        waitABit(15, environment)
-        examineLineageResults(reportId, fourReceivers, fakeItemCount)
-    }
-
-    private fun doStracBasicTest(
-        environment: TestingEnvironment
-    ) {
-        val fakeItemCount = 100
-        echo("StracBasic Test: sending Strac data to: ${environment.endPoint} using ${redoxReceiver.name}")
-        val file = createFakeFile(
-            metadata,
-            stracSender,
-            fakeItemCount,
-            receivingStates,
-            redoxReceiver.name,
-            dir,
-        )
-        echo("Created datafile $file")
-        // Now send it to the Hub.
-        val (responseCode, json) = postReportFile(environment, file, org.name, stracSender.name, key)
-        echo("Response to POST: $responseCode")
-        echo(json)
-        if (responseCode != HttpURLConnection.HTTP_CREATED) {
-            bad("***StracBasic Test FAILED***:  response code $responseCode")
-            return
-        }
-        val tree = jacksonObjectMapper().readTree(json)
-        val reportId = ReportId.fromString(tree["id"].textValue())
-        echo("Id of submitted report: $reportId")
-        waitABit(15, environment)
-        examineLineageResults(reportId, listOf(redoxReceiver), fakeItemCount)
-    }
-
-    private fun doHugeTest(environment: TestingEnvironment) {
-        val fakeItemCount = 10000
-        echo("Attempting to send $fakeItemCount items to ${environment.endPoint}. This is terrapin slow.")
-        val file = createFakeFile(
-            metadata,
-            simpleRepSender,
-            fakeItemCount,
-            receivingStates,
-            csvReceiver.name,
-            dir,
-            Report.Format.CSV
-        )
-        echo("Created datafile $file")
-        // Now send it to the Hub.
-        val (responseCode, json) = postReportFile(environment, file, org.name, simpleRepSender.name, key)
-        echo("Response to POST: $responseCode")
-        echo(json)
-        if (responseCode != HttpURLConnection.HTTP_CREATED) {
-            bad("***EndToEnd Test FAILED***:  response code $responseCode")
-            return
-        }
-        val tree = jacksonObjectMapper().readTree(json)
-        val reportId = ReportId.fromString(tree["id"].textValue())
-        echo("Id of submitted report: $reportId")
-        waitABit(20, environment)
-        examineLineageResults(reportId, listOf(csvReceiver), fakeItemCount)
-    }
-
-    private fun doTooManyItemsTest(environment: TestingEnvironment) {
-        val fakeItemCount = REPORT_MAX_ITEMS + 1
-        echo("Attempting to send $fakeItemCount items to ${environment.endPoint}. This is slllooooowww.")
-        val file = createFakeFile(
-            metadata,
-            simpleRepSender,
-            fakeItemCount,
-            receivingStates,
-            csvReceiver.name,
-            dir,
-            Report.Format.CSV
-        )
-        echo("Created datafile $file")
-        // Now send it to the Hub.
-        val (responseCode, json) = postReportFile(environment, file, org.name, simpleRepSender.name, key)
-        echo("Response to POST: $responseCode")
-        echo(json)
-        try {
-            val tree = jacksonObjectMapper().readTree(json)
-            val firstError = ((tree["errors"] as ArrayNode)[0]) as ObjectNode
-            if (firstError["details"].textValue().contains("rows")) {
-                good("Too Big Test passed.")
-            } else {
-                bad("***Too Big Test Test FAILED***: Did not find the error")
-            }
-        } catch (e: Exception) {
-            bad("***Too Big Test Test FAILED***: Unable to find the expected error message")
-        }
-    }
-
-    private fun doTooManyColumnsTest(
-        environment: TestingEnvironment
-    ) {
-        echo("Testing a file with too many columns.")
-        val file = File("./src/test/csv_test_files/input/too-many-columns.csv")
-        if (!file.exists()) {
-            error("Unable to find file ${file.absolutePath} to do toomanycols test")
-        }
-        val sendingOrg = settings.findOrganization(orgName)
-            ?: error("Unable to find org $orgName in metadata")
-        val sender = settings.findSender(orgName + "." + simpleReportSenderName)
-            ?: error("Unable to find sender $simpleReportSenderName for organization ${sendingOrg.name}")
-        // Now send it to the Hub.
-        val (responseCode, json) = postReportFile(environment, file, sendingOrg.name, sender.name, key)
-        echo("Response to POST: $responseCode")
-        echo(json)
-        try {
-            val tree = jacksonObjectMapper().readTree(json)
-            val firstError = ((tree["errors"] as ArrayNode)[0]) as ObjectNode
-            if (firstError["details"].textValue().contains("columns")) {
-                good("Test passed: Too many columns test.")
-            } else {
-                bad("***Too Many Columns Test FAILED***:  did not find the error.")
-            }
-        } catch (e: Exception) {
-            bad("***Too Many Columns Test FAILED***: Unable to find the expected error message")
-        }
-    }
-
+    // ping
     private fun doCheckConnections(
-        environment: TestingEnvironment
+        environment: ReportStreamEnv
     ) {
         echo("CheckConnections of ${environment.endPoint}")
-        val (responseCode, json) = postReportBytes(
+        val (responseCode, json) = HttpUtilities.postReportBytes(
             environment,
             "x".toByteArray(),
             orgName,
@@ -363,12 +180,13 @@ Examples:
         }
     }
 
+    // end2end
     private fun doEndToEndTest(
-        environment: TestingEnvironment
+        environment: ReportStreamEnv
     ) {
         echo("EndToEndTest of: ${environment.endPoint}")
         val fakeItemCount = 20 // hack:  you need use a multiple of # of targetCounties
-        val file = createFakeFile(
+        val file = FileUtilities.createFakeFile(
             metadata,
             simpleRepSender,
             fakeItemCount,
@@ -378,7 +196,7 @@ Examples:
         )
         echo("Created datafile $file")
         // Now send it to the Hub.
-        val (responseCode, json) = postReportFile(environment, file, org.name, simpleRepSender.name, key)
+        val (responseCode, json) = HttpUtilities.postReportFile(environment, file, org.name, simpleRepSender.name, key)
         echo("Response to POST: $responseCode")
         echo(json)
         if (responseCode != HttpURLConnection.HTTP_CREATED) {
@@ -394,6 +212,183 @@ Examples:
         } catch (e: NullPointerException) {
             bad("***End to End Test FAILED***: Unable to properly parse response json")
         }
+    }
+
+    private fun doMergeTest(environment: ReportStreamEnv) {
+        val fakeItemCount = 20 // hack:  you need use a multiple of # of targetCounties
+        echo("Merge test of: ${environment.endPoint} across $fourTargetCounties")
+        val file = FileUtilities.createFakeFile(
+            metadata,
+            simpleRepSender,
+            fakeItemCount,
+            receivingStates,
+            fourTargetCounties,
+            dir,
+        )
+        echo("Created datafile $file")
+        // Now send it to the Hub over and over
+        val reportIds = (1..5).map {
+            val (responseCode, json) =
+                HttpUtilities.postReportFile(environment, file, org.name, simpleRepSender.name, key)
+            echo("Response to POST: $responseCode")
+            if (responseCode != HttpURLConnection.HTTP_CREATED) {
+                bad("***Merge Test FAILED***:  response code $responseCode")
+                return
+            }
+            val tree = jacksonObjectMapper().readTree(json)
+            val reportId = ReportId.fromString(tree["id"].textValue())
+            echo("Id of submitted report: $reportId")
+            reportId
+        }
+        waitABit(30, environment)
+        examineMergeResults(reportIds[0], fourReceivers, fakeItemCount, 5)
+    }
+
+    private fun doHugeTest(environment: ReportStreamEnv) {
+        val fakeItemCount = 10000
+        echo("Attempting to send $fakeItemCount items to ${environment.endPoint}. This is terrapin slow.")
+        val file = FileUtilities.createFakeFile(
+            metadata,
+            simpleRepSender,
+            fakeItemCount,
+            receivingStates,
+            csvReceiver.name,
+            dir,
+            Report.Format.CSV
+        )
+        echo("Created datafile $file")
+        // Now send it to the Hub.
+        val (responseCode, json) = HttpUtilities.postReportFile(environment, file, org.name, simpleRepSender.name, key)
+        echo("Response to POST: $responseCode")
+        echo(json)
+        if (responseCode != HttpURLConnection.HTTP_CREATED) {
+            bad("***EndToEnd Test FAILED***:  response code $responseCode")
+            return
+        }
+        val tree = jacksonObjectMapper().readTree(json)
+        val reportId = ReportId.fromString(tree["id"].textValue())
+        echo("Id of submitted report: $reportId")
+        waitABit(20, environment)
+        examineLineageResults(reportId, listOf(csvReceiver), fakeItemCount)
+    }
+
+    private fun doTooManyItemsTest(environment: ReportStreamEnv) {
+        val fakeItemCount = REPORT_MAX_ITEMS + 1
+        echo("Attempting to send $fakeItemCount items to ${environment.endPoint}. This is slllooooowww.")
+        val file = FileUtilities.createFakeFile(
+            metadata,
+            simpleRepSender,
+            fakeItemCount,
+            receivingStates,
+            csvReceiver.name,
+            dir,
+            Report.Format.CSV
+        )
+        echo("Created datafile $file")
+        // Now send it to the Hub.
+        val (responseCode, json) = HttpUtilities.postReportFile(environment, file, org.name, simpleRepSender.name, key)
+        echo("Response to POST: $responseCode")
+        echo(json)
+        try {
+            val tree = jacksonObjectMapper().readTree(json)
+            val firstError = ((tree["errors"] as ArrayNode)[0]) as ObjectNode
+            if (firstError["details"].textValue().contains("rows")) {
+                good("Too Big Test passed.")
+            } else {
+                bad("***Too Big Test Test FAILED***: Did not find the error")
+            }
+        } catch (e: Exception) {
+            bad("***Too Big Test Test FAILED***: Unable to find the expected error message")
+        }
+    }
+
+    private fun doTooManyColumnsTest(
+        environment: ReportStreamEnv
+    ) {
+        echo("Testing a file with too many columns.")
+        val file = File("./src/test/csv_test_files/input/too-many-columns.csv")
+        if (!file.exists()) {
+            error("Unable to find file ${file.absolutePath} to do toomanycols test")
+        }
+        val sendingOrg = settings.findOrganization(orgName)
+            ?: error("Unable to find org $orgName in metadata")
+        val sender = settings.findSender(orgName + "." + simpleReportSenderName)
+            ?: error("Unable to find sender $simpleReportSenderName for organization ${sendingOrg.name}")
+        // Now send it to the Hub.
+        val (responseCode, json) = HttpUtilities.postReportFile(environment, file, sendingOrg.name, sender.name, key)
+        echo("Response to POST: $responseCode")
+        echo(json)
+        try {
+            val tree = jacksonObjectMapper().readTree(json)
+            val firstError = ((tree["errors"] as ArrayNode)[0]) as ObjectNode
+            if (firstError["details"].textValue().contains("columns")) {
+                good("Test passed: Too many columns test.")
+            } else {
+                bad("***Too Many Columns Test FAILED***:  did not find the error.")
+            }
+        } catch (e: Exception) {
+            bad("***Too Many Columns Test FAILED***: Unable to find the expected error message")
+        }
+    }
+
+    // Strac format (only) to Redox format (only)
+    private fun doStracBasicTest(
+        environment: ReportStreamEnv
+    ) {
+        val fakeItemCount = 100
+        echo("StracBasic Test: sending Strac data to: ${environment.endPoint} using ${redoxReceiver.name}")
+        val file = FileUtilities.createFakeFile(
+            metadata,
+            stracSender,
+            fakeItemCount,
+            receivingStates,
+            redoxReceiver.name,
+            dir,
+        )
+        echo("Created datafile $file")
+        // Now send it to the Hub.
+        val (responseCode, json) = HttpUtilities.postReportFile(environment, file, org.name, stracSender.name, key)
+        echo("Response to POST: $responseCode")
+        echo(json)
+        if (responseCode != HttpURLConnection.HTTP_CREATED) {
+            bad("***StracBasic Test FAILED***:  response code $responseCode")
+            return
+        }
+        val tree = jacksonObjectMapper().readTree(json)
+        val reportId = ReportId.fromString(tree["id"].textValue())
+        echo("Id of submitted report: $reportId")
+        waitABit(15, environment)
+        examineLineageResults(reportId, listOf(redoxReceiver), fakeItemCount)
+    }
+
+    // strac format to all output formats
+    private fun doStracTest(
+        environment: ReportStreamEnv
+    ) {
+        echo("Test sending Strac data to: ${environment.endPoint} across $fourTargetCounties")
+        val fakeItemCount = 20 // hack:  you need use a multiple of # of targetCounties
+        val file = FileUtilities.createFakeFile(
+            metadata,
+            stracSender,
+            fakeItemCount,
+            receivingStates,
+            fourTargetCounties,
+            dir,
+        )
+        echo("Created datafile $file")
+        // Now send it to the Hub.
+        val (responseCode, json) = HttpUtilities.postReportFile(environment, file, org.name, stracSender.name, key)
+        echo("Response to POST: $responseCode")
+        echo(json)
+        if (responseCode != HttpURLConnection.HTTP_CREATED) {
+            bad("**Strac Test FAILED***:  response code $responseCode")
+            return
+        }
+        val tree = jacksonObjectMapper().readTree(json)
+        val reportId = ReportId.fromString(tree["id"].textValue())
+        echo("Id of submitted report: $reportId")
+        waitABit(15, environment)
+        examineLineageResults(reportId, fourReceivers, fakeItemCount)
     }
 
     fun examineLineageResults(
@@ -454,84 +449,54 @@ Examples:
     }
 
     companion object {
+        val metadata = Metadata(Metadata.defaultMetadataDirectory)
+        val settings = FileSettings(FileSettings.defaultSettingsDirectory)
+
+        // Here is test setup of organization, senders, and receivers.   All static.
+        val orgName = "ignore"
+        val org = settings.findOrganization(orgName)
+            ?: error("Unable to find org $orgName in metadata")
+        val receivingStates = "IG"
+
+        val simpleReportSenderName = "ignore-simple-report"
+        val simpleRepSender = settings.findSender("$orgName.$simpleReportSenderName")
+            ?: error("Unable to find sender $simpleReportSenderName for organization ${org.name}")
+
+        val stracSenderName = "ignore-strac"
+        val stracSender = settings.findSender("$orgName.$stracSenderName")
+            ?: error("Unable to find sender $stracSenderName for organization ${org.name}")
+
+        val fourReceivers = settings.receivers.filter { it.organizationName == orgName }
+        val csvReceiver = fourReceivers.filter { it.name == "CSV" }[0]
+        val redoxReceiver = fourReceivers.filter { it.name == "REDOX" }[0]
+        val fourTargetCounties = fourReceivers.map { it.name }.joinToString(",")
 
         const val ANSI_RESET = "\u001B[0m"
         const val ANSI_BLACK = "\u001B[30m"
         const val ANSI_RED = "\u001B[31m"
         const val ANSI_GREEN = "\u001B[32m"
         const val ANSI_BLUE = "\u001B[34m"
-
-        val receivingStates = "IG"
-        val orgName = "ignore"
-        val simpleReportSenderName = "ignore-simple-report"
-        val stracSenderName = "ignore-strac"
-        val metadata = Metadata(Metadata.defaultMetadataDirectory)
-        val settings = FileSettings(FileSettings.defaultSettingsDirectory)
-        val org = settings.findOrganization(orgName)
-            ?: error("Unable to find org $orgName in metadata")
-        val simpleRepSender = settings.findSender(orgName + "." + simpleReportSenderName)
-            ?: error("Unable to find sender $simpleReportSenderName for organization ${org.name}")
-        val stracSender = settings.findSender(orgName + "." + simpleReportSenderName)
-            ?: error("Unable to find sender $stracSenderName for organization ${org.name}")
-        val fourReceivers = settings.receivers.filter { it.organizationName == orgName }
-        val csvReceiver = fourReceivers.filter { it.name == "CSV" }[0]
-        val redoxReceiver = fourReceivers.filter { it.name == "REDOX" }[0]
-        val fourTargetCounties = fourReceivers.map { it.name }.joinToString(",")
-
+        const val ANSI_CYAN = "\u001B[36m"
         fun good(msg: String) {
             echo(ANSI_GREEN + msg + ANSI_RESET)
         }
-
         fun bad(msg: String) {
             echo(ANSI_RED + msg + ANSI_RESET)
         }
-
-        fun createFakeFile(
-            metadata: Metadata,
-            sender: Sender,
-            count: Int,
-            targetStates: String? = null,
-            targetCounties: String? = null,
-            directory: String = ".",
-            format: Report.Format = Report.Format.CSV,
-        ): File {
-            val report = createFakeReport(
-                metadata,
-                sender,
-                count,
-                targetStates,
-                targetCounties,
-            )
-            return writeReportToFile(report, format, metadata, directory, null)
-        }
-
-        fun createFakeReport(
-            metadata: Metadata,
-            sender: Sender,
-            count: Int,
-            targetStates: String? = null,
-            targetCounties: String? = null,
-        ): Report {
-            return FakeReport(metadata).build(
-                metadata.findSchema(sender.schemaName)
-                    ?: error("Unable to find schema ${sender.schemaName}"),
-                count,
-                FileSource("fake"),
-                targetStates,
-                targetCounties,
-            )
+        fun ugly(msg: String) {
+            echo(ANSI_CYAN + msg + ANSI_RESET)
         }
 
         /**
          * A hack attempt to wait enough time, but not too long, for Hub to finish.
          * This assumes the batch/send executes on the minute boundary.
          */
-        fun waitABit(plusSecs: Int, env: TestReportStream.TestingEnvironment) {
+        fun waitABit(plusSecs: Int, env: ReportStreamEnv) {
             // seconds elapsed so far in this minute
             val secsElapsed = OffsetDateTime.now().second % 60
             // Wait until the top of the next minute, and pluSecs more, for 'batch', and 'send' to finish.
             var waitSecs = 60 - secsElapsed + plusSecs
-            if (secsElapsed > (60 - plusSecs) || env != TestingEnvironment.LOCAL) {
+            if (secsElapsed > (60 - plusSecs) || env != ReportStreamEnv.LOCAL) {
                 // Uh oh, we are close to the top of the minute *now*, so 'receive' might not finish in time.
                 // Or, we are in Test or Staging, which don't execute on the top of the minute.
                 waitSecs += 60
@@ -542,139 +507,6 @@ Examples:
                 print(".")
             }
             println()
-        }
-
-        /**
-         * A generic function to POST a Prime Data Hub report File to a particular Prime Data Hub Environment,
-         * as if from sendingOrgName.sendingOrgClientName.
-         * Returns Pair(Http response code, json response text)
-         */
-        fun postReportFile(
-            environment: TestingEnvironment,
-            file: File,
-            sendingOrgName: String,
-            sendingOrgClientName: String? = null,
-            key: String? = null,
-            option: ReportFunction.Options ? = null
-        ): Pair<Int, String> {
-            if (!file.exists()) error("Unable to find file ${file.absolutePath}")
-            return postReportBytes(environment, file.readBytes(), sendingOrgName, sendingOrgClientName, key, option)
-        }
-
-        /**
-         * A generic function to POST data to a particular Prime Data Hub Environment,
-         * as if from sendingOrgName.sendingOrgClientName.
-         * Returns Pair(Http response code, json response text)
-         */
-        private fun postReportBytes(
-            environment: TestingEnvironment,
-            bytes: ByteArray,
-            sendingOrgName: String,
-            sendingOrgClientName: String?,
-            key: String?,
-            option: ReportFunction.Options?
-        ): Pair<Int, String> {
-            val headers = mutableListOf<Pair<String, String>>()
-            headers.add("Content-Type" to "text/csv")
-            val clientStr = sendingOrgName + if (sendingOrgClientName != null) ".$sendingOrgClientName" else ""
-            headers.add("client" to clientStr)
-            if (key == null && environment == TestingEnvironment.TEST) error("key is required for Test environment")
-            if (key != null)
-                headers.add("x-functions-key" to key)
-            val url = environment.endPoint + if (option != null) "?option=$option" else ""
-            return postHttp(url, bytes, headers)
-        }
-
-        /**
-         * A generic function that posts data to a URL <address>.
-         * Returns the HTTP response code and the text of the response
-         */
-        fun postHttp(urlStr: String, bytes: ByteArray, headers: List<Pair<String, String>>? = null): Pair<Int, String> {
-            val urlObj = URL(urlStr)
-            with(urlObj.openConnection() as HttpURLConnection) {
-                requestMethod = "POST"
-                doOutput = true
-                doInput = true
-                headers?.forEach {
-                    addRequestProperty(it.first, it.second)
-                }
-                outputStream.use {
-                    it.write(bytes)
-                }
-                val response = try {
-                    inputStream.bufferedReader().readText()
-                } catch (e: IOException) {
-                    return responseCode to responseMessage
-                }
-                return responseCode to response
-            }
-        }
-
-        // todo this belongs in some other utility class.  simple version of what's in main.
-        fun writeReportsToFile(
-            reports: List<Pair<Report, Report.Format>>,
-            metadata: Metadata,
-            outputDir: String?,
-            outputFileName: String?,
-        ) {
-            if (outputDir == null && outputFileName == null) return
-
-            if (reports.isNotEmpty()) {
-                echo("Creating these files:")
-            }
-            reports
-                .flatMap { (report, format) ->
-                    // Some report formats only support one result per file
-                    if (format.isSingleItemFormat()) {
-                        val splitReports = report.split()
-                        splitReports.map { Pair(it, format) }
-                    } else {
-                        listOf(Pair(report, format))
-                    }
-                }.forEach { (report, format) ->
-                    var outputFile = writeReportToFile(report, format, metadata, outputDir, outputFileName)
-                    echo(outputFile.absolutePath)
-                }
-        }
-
-        // todo this belongs in some other utility class.  simple version of what's in main.
-        fun writeReportToFile(
-            report: Report,
-            format: Report.Format,
-            metadata: Metadata,
-            outputDir: String?,
-            outputFileName: String?,
-        ): File {
-            val outputFile = if (outputFileName != null) {
-                File(outputFileName)
-            } else {
-                val fileName = Report.formFilename(
-                    report.id,
-                    report.schema.baseName,
-                    format,
-                    report.createdDateTime,
-                    useAphlFormat = false,
-                    report.destination?.translation?.receivingOrganization,
-                )
-                File(outputDir ?: ".", fileName)
-            }
-            if (!outputFile.exists()) {
-                outputFile.createNewFile()
-            }
-
-            val csvSerializer = CsvSerializer(metadata)
-            val hl7Serializer = Hl7Serializer(metadata)
-            val redoxSerializer = RedoxSerializer(metadata)
-            outputFile.outputStream().use {
-                when (format) {
-                    Report.Format.INTERNAL -> csvSerializer.writeInternal(report, it)
-                    Report.Format.CSV -> csvSerializer.write(report, it)
-                    Report.Format.HL7 -> hl7Serializer.write(report, it)
-                    Report.Format.HL7_BATCH -> hl7Serializer.writeBatch(report, it)
-                    Report.Format.REDOX -> redoxSerializer.write(report, it)
-                }
-            }
-            return outputFile
         }
 
         fun itemLineageCountQuery(
@@ -694,28 +526,28 @@ Examples:
               (select item_descendants(?)) """
             return ctx.fetchOne(sql, receivingOrgSvc, action, reportId)?.into(Int::class.java)
         }
-    }
 
-    fun reportLineageCountQuery(
-        txn: DataAccessTransaction,
-        reportId: ReportId,
-        receivingOrgSvc: String,
-        action: TaskAction,
-    ): Int? {
-        val ctx = DSL.using(txn)
-        val sql = """select sum(item_count)
+        fun reportLineageCountQuery(
+            txn: DataAccessTransaction,
+            reportId: ReportId,
+            receivingOrgSvc: String,
+            action: TaskAction,
+        ): Int? {
+            val ctx = DSL.using(txn)
+            val sql = """select sum(item_count)
               from report_file as RF
               join action as A on A.action_id = RF.action_id
               where RF.receiving_org_svc = ? 
               and A.action_name = ? 
               and RF.report_id in
               (select report_descendants(?)) """
-        return ctx.fetchOne(sql, receivingOrgSvc, action, reportId)?.into(Int::class.java)
+            return ctx.fetchOne(sql, receivingOrgSvc, action, reportId)?.into(Int::class.java)
+        }
     }
 }
 
 /**
- * This was my attempt to jooqify the query.
+ * This was my last attempt to jooqify the above itemLineage query.
  *  Remaining issue is the call to (select from item_descendants(report_id) part.
  *  I'm sure there's a correct jooq way to call a stored fn that returns a set, but couldn't find it.
  *  Could also treat just that part as "plain SQL", but I didn't see the difference
